chore: fix clippy warnings across workspace
- Fix clippy lints across D3D11/D3D9, emulator, virtio, firmware, wasm, and tests
- Replace manual alignment/modulo checks with `is_multiple_of`/`div_ceil` and remove identity ops
- Tidy helper APIs and test harnesses (argument structs, slice params, defaults) to keep `-D warnings` clean

+++ crates/aero-cpu-core/src/interp/sse.rs
@@ -61,10 +61,10 @@ fn cvt_i64_to_f32(cpu: &mut CpuState, src: i64) -> f32 {
     let shift = (msb as i32) - 23;
 
     let (mut mantissa_full, rem) = if shift <= 0 {
-        ((mag << (-shift as u32)) as u64, 0u64)
+        (mag << (-shift as u32), 0u64)
     } else {
         let shift = shift as u32;
-        ((mag >> shift) as u64, mag & ((1u64 << shift) - 1))
+        (mag >> shift, mag & ((1u64 << shift) - 1))
     };
 
     if rem != 0 {
